requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
lightgbm>=4.0
xgboost>=2.0
scipy>=1.11
yfinance>=0.2.31
pyyaml>=6.0
joblib>=1.3
matplotlib>=3.7
pyarrow>=14.0
shap>=0.43
pytest>=7.0
